Make googletest dependency optional (#3949)
Dave reported an issue while rebuilding 2.13.2 from the source package
using koji. Because that build is done without network access, the build
failed with a missing googletest dependency due to
https://github.com/cvmfs/cvmfs/pull/3874

We've added a packaging patch to include googletest in the src package
in https://github.com/cvmfs/cvmfs/tree/cvmfs-2.13.2-2 , but there's
really no good reason for our unittest library to be a required
dependency. This PR makes it so that googletest is only needed when the
unittests are actually built, by adding a forwarding header
`duplex_testing.h` that is included instead of `gtest_prod.h`. It also
makes the cvmfs rpm spec file skip the unittest package by default. It
can still be built with `rpmbuild --define="build_testing 1"`. I don't
really get the need to package the unittests though, and might remove
this alltogether in the future ( this PR already removes the unittests
from debian packaging ). I find that it's more important to have the
option to run the unittests during the package build, failing early
before even creating the package. Before we can enable this by default,
the T_Namespace tests should be looked at or disabled though, because
they make trouble when testing in a container.

Just in case, it also allows for the googletest source to be downloaded
to the `externals` directory before cmake is invoked. This can serve as
a template for the actually required, vendored dependencies that don't
have to be checked into the repo with a CMake configured that way.