[util] Complete redesign of the interface. See details bellow
Previously the interface was designed for exchanging strings
(std::string specifically).

To signal the end of transmission a specific sequence had to be read
during read.

Now any compact/contiguous type can be exchanged between the client and
the server.

In multiple exchanges it is the responsibility of the sender to inform
the receiver about the number of data, and of the receiver to poll for
the correct amount of data.